fix: Fixed several local file system api
