+++ src/components/ChangeFilePathDialog.tsx
@@ -7,6 +7,7 @@ import {
   TextField,
 } from "@material-ui/core";
 import { TabNode } from "flexlayout-react";
+import Noty from "noty";
 import React, { useCallback, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { CrossnoteContainer } from "../containers/crossnote";
