+++ src/i18n/lang/zhTW.ts
@@ -196,6 +196,7 @@ export const zhTW = {
     "general/unpin-the-note": "取消置頂這個筆記",
     "general/add-to-quick-access": "添加到快速訪問",
     "general/remove-from-quick-access": "從快速訪問中移除",
+    "general/open-folder": "打開文件夾",
 
     // Error
     "error/authentication-failed": "認証失敗",
